Revert "Cleanup + new feature"
