--- .eslintrc
@@ -1,11 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "extends": "airbnb",
-  "rules": {
-    "react/jsx-filename-extension": 0,
-    "import/prefer-default-export": 0
-  },
-  "env": {
-    "browser": true
-  }
-}
