+++ package.json
@@ -3,23 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.17.1",
-    "prop-types": "^15.6.0",
-    "react": "16.1.1",
-    "react-dom": "16.1.1",
+    "axios": "^0.15.3",
+    "react": "15.4.2",
+    "react-dom": "15.4.2",
     "react-redux": "^5.0.3",
     "redux": "^3.6.0",
-    "redux-saga": "^0.16.0"
+    "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "eslint": "^4.9.0",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.4.0",
-    "json-server": "^0.12.1",
-    "react-scripts": "1.0.17",
-    "redux-devtools-extension": "^2.13.2"
+    "json-server": "^0.9.5",
+    "react-scripts": "0.9.1"
   },
   "scripts": {
     "api": "json-server --watch api/db.json --port 3001",
