+++ src/App.js
@@ -1,20 +1,48 @@
-import React from 'react';
+import React, { Component, PropTypes } from 'react';
+import { connect } from 'react-redux';
 import logo from './logo.svg';
 import './App.css';
 
-import JediList from './components/JediList';
-import JediForm from './components/JediForm';
+import { fetchJedi } from './action';
 
-const App = () =>
-  (
-    <div className="App">
-      <div className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <h2>Welcome to React</h2>
+function mapStateToProps(state) {
+  return {
+    jedi: state.jedi,
+  };
+}
+
+class App extends Component {
+  componentWillMount() {
+    this.fetchJedi();
+  }
+
+  fetchJedi() {
+    this.props.dispatch(fetchJedi());
+  }
+
+  render() {
+    const { jedi } = this.props;
+
+    return (
+      <div className="App">
+        <div className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+          <h2>Welcome to React</h2>
+        </div>
+        {jedi.map((jedi, index) => (
+          <div key={index}>
+            Jedi: id: {jedi.id} name: {jedi.name}
+          </div>
+        ))}
       </div>
-      <JediList />
-      <JediForm />
-    </div>
-  );
+    );
+  }
+}
 
-export default App;
+App.propTypes = {
+  jedi: PropTypes.array,
+};
+
+export default connect(
+  mapStateToProps,
+)(App);
