+++ src/action.js
@@ -1,0 +1,17 @@
+/**
+ * Created by thomashourlier on 2/26/17.
+ */
+
+import axios from 'axios';
+
+export function fetchJedi() {
+  return (dispatch) => {
+    axios.get('http://localhost:3001/jedi')
+      .then((res) => {
+        dispatch({
+          type: 'FETCH_FINISH',
+          payload: res.data,
+        });
+      })
+  }
+}
