--- src/components/JediForm/JediForm.container.js
@@ -1,11 +1,0 @@
-// @flow
-
-import { connect } from 'react-redux';
-import { addJedi } from '../../redux/actions';
-import JediForm from './JediForm.component';
-
-const mapDispatchToProps = {
-  addJedi,
-};
-
-export default connect(null, mapDispatchToProps)(JediForm);
