--- src/components/JediForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './JediForm.container';
