--- src/components/JediList/index.js
@@ -1,1 +1,0 @@
-export { default } from './JediList.container';
