+++ src/index.js
@@ -4,11 +4,11 @@ import { Provider } from 'react-redux';
 import App from './App';
 import './index.css';
 
-import store from './redux/store';
+import store from './store';
 
 ReactDOM.render(
   <Provider store={store}>
     <App />
   </Provider>,
-  document.getElementById('root'),
+  document.getElementById('root')
 );
