--- src/redux/constants.js
@@ -1,4 +1,0 @@
-export const ADD_NEW_JEDI = 'ADD_NEW_JEDI';
-export const FETCH_JEDI = 'FETCH_JEDI';
-export const JEDI_ERROR = 'JEDI_ERROR';
-export const JEDI_RECEIVED = 'JEDI_RECEIVED';
